api/requirements.txt
fastapi
uvicorn
python-dotenv
azure-ai-language-conversations
azure-ai-textanalytics
azure-core
